movies_async_api/tests/functional/requirements.txt
backoff==1.10.0
elasticsearch==7.10.0
pytest==6.1.2
python-decouple==3.3
redis==3.5.3
requests==2.25.0
